add Speed-Tests file
The goal is to compare BE backed json functions to this project to the JSON
module project which converts to Let Notation.
Currently, both sets of custom functions can walk any array by detecting if the
value is an object or array, but the JSON module cannot, so the current testing
implementation isn't scalable.
I think it's good enough to use as a basis for some code refactoring, though.