move parseWhitespace code to a support function
This removed the last references to jsonParseNext and jsonDiscardResult, so I
deleted them. This commit completes the consolidation of the parsing functions
into 3 support functions.

Also modified to use chunking method, which reduces recursion in this code. This
change allows the jsonGet large mixed data test to process ~4 times more data.