closes #34; timestamps not encoded properly
NOTE: timestamps, dates, and times are converted to numbers in FileMaker's
own internal number format. This is useful for parsing within FileMaker,
but not useful for interacting with other services that expect a standard
format like YYYY-MM-DD.
Currently, you must manually convert timestamps, dates, and times to
whatever standard format you need, if sending the JSON to a different
system that expects a particular format.